refactor(caldav): move calendar-query REPORT overrides to dedicated Report class
Isolate the ESN specific calendar-query REPORT logic in a dedicated
lib/CalDAV/Report.php class, keeping only the thin override entrypoint in
the Plugin, which delegates to it. This collocates the report overrides
and keeps them separated from the rest of the MongoDB CalDAV plugin, as
suggested in review.
